[Win32] Correct DPI context awareness value retrieval
The DPI_AWARENESS_CONTEXT values are currently defined as constants, but
they are actually defined as typedefed handles instead of compile time
constants in Windows' windef.h. So the constants defined in SWT may not
fit the actual values as used by the OS. This currently seems to be the
case for the System awareness value, which does not fit the OS value.

This change adapts the definition of the values to retrieve the actual
current values from the OS instead of hardcoding them in constants.

+++ bundles/org.eclipse.swt/Eclipse SWT/win32/org/eclipse/swt/internal/Win32DPIUtils.java
@@ -42,7 +42,7 @@ public class Win32DPIUtils {
 	 */
 	public static final String CUSTOM_DPI_AWARENESS_PROPERTY = "org.eclipse.swt.internal.win32.dpiAwareness";
 
-	private static int customDpiAwareness = -1;
+	private static long customDpiAwareness = -1;
 
 	public static void initializeCustomDpiAwareness() {
 		String customDpiAwareness = System.getProperty(CUSTOM_DPI_AWARENESS_PROPERTY);
@@ -63,7 +63,7 @@ public class Win32DPIUtils {
 		}
 	}
 
-	public static boolean setDPIAwareness(int desiredDpiAwareness) {
+	public static boolean setDPIAwareness(long desiredDpiAwareness) {
 		if (desiredDpiAwareness == OS.GetThreadDpiAwarenessContext()) {
 			return true;
 		}
